+++ hydromt_wflow/wflow.py
@@ -1546,6 +1546,9 @@ class WflowModel(Model):
         pet_method: str = "debruin",
         press_correction: bool = True,
         temp_correction: bool = True,
+        wind_correction: bool = True,
+        wind_altitude: int = 10,
+        reproj_method: str = "nearest_index",
         dem_forcing_fn: str = "era5_orography",
         skip_pet: str = False,
         chunksize: Optional[int] = None,
@@ -1562,15 +1565,18 @@ class WflowModel(Model):
         ----------
         temp_pet_fn : str, optional
             Name or path of data source with variables to calculate temperature and reference evapotranspiration,
-            see data/forcing_sources.yml, by default 'era5'.
+            see data/forcing_sources.yml, by default 'era5_daily_zarr'.
 
             * Required variable for temperature: ['temp']
 
             * Required variables for De Bruin reference evapotranspiration: ['temp', 'press_msl', 'kin', 'kout']
 
             * Required variables for Makkink reference evapotranspiration: ['temp', 'press_msl', 'kin']
-        pet_method : {'debruin', 'makkink'}, optional
+
+            * Required variables for daily Penman-Monteith reference evapotranspiration: either ['temp', 'temp_min', 'temp_max', 'wind', 'rh', 'kin'] for 'penman-monteith_rh_simple' or ['temp', 'temp_min', 'temp_max', 'temp_dew', 'wind', 'kin', 'press_msl', "wind10_u", "wind10_v"] for 'penman-monteith_tdew' (these are the variables available in ERA5)
+        pet_method : {'debruin', 'makkink', 'penman-monteith_rh_simple', 'penman-monteith_tdew'}, optional
             Reference evapotranspiration method, by default 'debruin'.
+            If penman-monteith is used, requires the installation of the pyet package.
         press_correction, temp_correction : bool, optional
              If True pressure, temperature are corrected using elevation lapse rate,
              by default False.
@@ -1599,8 +1605,25 @@ class WflowModel(Model):
                 variables += ["press_msl", "kin", "kout"]
             elif pet_method == "makkink":
                 variables += ["press_msl", "kin"]
+            elif pet_method == "penman-monteith_rh_simple":
+                variables += ["temp_min", "temp_max", "wind", "rh", "kin"]
+            elif pet_method == "penman-monteith_tdew":
+                variables += [
+                    "temp_min",
+                    "temp_max",
+                    "wind10_u",
+                    "wind10_v",
+                    "temp_dew",
+                    "kin",
+                    "press_msl",
+                ]
             else:
-                methods = ["debruin", "makking"]
+                methods = [
+                    "debruin",
+                    "makking",
+                    "penman-monteith_rh_simple",
+                    "penman-monteith_tdew",
+                ]
                 ValueError(f"Unknown pet method {pet_method}, select from {methods}")
 
         ds = self.data_catalog.get_rasterdataset(
@@ -1633,13 +1656,43 @@ class WflowModel(Model):
             **kwargs,
         )
 
+        if (
+            "penman-monteith" in pet_method
+        ):  # also downscaled temp_min and temp_max for Penman needed
+            temp_max_in = hydromt.workflows.forcing.temp(
+                ds["temp_max"],
+                dem_model=self.staticmaps[self._MAPS["elevtn"]],
+                dem_forcing=dem_forcing,
+                lapse_correction=temp_correction,
+                logger=self.logger,
+                freq=None,  # resample time after pet workflow
+                **kwargs,
+            )
+            temp_max_in.name = "temp_max"
+
+            temp_min_in = hydromt.workflows.forcing.temp(
+                ds["temp_min"],
+                dem_model=self.staticmaps[self._MAPS["elevtn"]],
+                dem_forcing=dem_forcing,
+                lapse_correction=temp_correction,
+                logger=self.logger,
+                freq=None,  # resample time after pet workflow
+                **kwargs,
+            )
+            temp_min_in.name = "temp_min"
+
+            temp_in = xr.merge([temp_in, temp_max_in, temp_min_in])
+
         if not skip_pet:
             pet_out = hydromt.workflows.forcing.pet(
                 ds[variables[1:]],
-                dem_model=self.staticmaps[self._MAPS["elevtn"]],
                 temp=temp_in,
+                dem_model=self.staticmaps[self._MAPS["elevtn"]],
                 method=pet_method,
                 press_correction=press_correction,
+                wind_correction=wind_correction,
+                wind_altitude=wind_altitude,
+                reproj_method=reproj_method,
                 freq=freq,
                 resample_kwargs=dict(label="right", closed="right"),
                 logger=self.logger,
@@ -1653,6 +1706,9 @@ class WflowModel(Model):
             pet_out.attrs.update(opt_attr)
             self.set_forcing(pet_out.where(mask), name="pet")
 
+        # make sure only temp is written to netcdf
+        if "penman-monteith" in pet_method:
+            temp_in = temp_in["temp"]
         # resample temp after pet workflow
         temp_out = hydromt.workflows.forcing.resample_time(
             temp_in,