+++ docs/changelog.rst
@@ -29,6 +29,9 @@ Fixed
 Deprecated
 ----------
 
+Added
+^^^^^
+- add options to calculate daily Penman-Monteith potential evaporation using the pyet package. Depending on the available variables, two options are defined ``penman-monteith_tdew`` (inputs: ['temp', 'temp_min', 'temp_max', 'wind_u', 'wind_v', 'temp_dew', 'kin', 'press_msl']) and ``penman-monteith_rh_simple`` (inputs: ['temp', 'temp_min', 'temp_max', 'wind', 'rh', 'kin']).
 
 v0.2.1 (22 November 2022)
 =========================